Initial metadata reading implementation
The overall goal is to build out a tool that will read a directory of
photos and sort them into a directory structure of <year>/<month>/<day>.
Many of my photos are CR2 raw photo files and so the first step of this
is tool is to be able to read the date out of a single one of these
photos.

The current implementation is pretty brittle, I inspected a number of my
photos using hexdump and found the byte offset that the date starts at
and am using that. There is likely some more reliable structural
context I can read but this is a start. I can do some high level
validation that the read data conforms to my expectations for the time
being.

Next steps include parsing out the year, month, day from the resulting
string, and either outputting the destination, next we can work on
moving, finally we want to do this concurrently for a whole directory of
image files. I will likely also eventually want to do the same for JPEG.
One complication of moving the files within this tool is that I will
likely want to do this within a git repo and so issuing git-mv commands
as opposed to regular file moves might be desirable.

TODO: Probably look into the Exif format to see if that is applicable to
the raw photos.