Somewhat more robust seeking algorithm
Rather than seeking to a specific offset and trying to parse whatever
was there as our date, I use a set of common bytes in the initial part
of the file as "markers" to seek through to find our way to the date.
This is a bit more reliable in that it now works for (most) of the JPGs
I've tried it with and CR2 files from both of the cameras that I have.

It is not totally robust in that it has the possibility of failing if
any of the bytes we used as "markers" are in other parts of the file
before the date. That said, at least in this case we should realize that
while seeking and bail out. I actually saw a few JPG files that have the
initial 'I' byte from the 'II*' sequency after some other 'I' byte.
We'll build this up to be even more robust for that sort of thing and
then maybe look at some specs to see what this all means in more detail.