Even more flexible finding of initial exif II* sequence
Instead of finding the first 'I' character, which was error prone due to
some JPG files having a single 'I' byte before the start of the exif
data, scan the first 16 bytes for the full 'II*' and then start
walking the file to get our date from that point.

The version of photosort successfully sorted all of the 1200-1300 photos
in a memory card dump from two different cannon photos with CR2 and JPG
file from both cameras in that directory. The previous version missed a
few JPGs due to the above issue.